Make `Audio` use the new phase vocoder operator
Also removed the old one
